backend/requirements_dev.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements_dev.in
#
asgiref==3.7.2
    # via
    #   -c requirements.txt
    #   django
build==1.0.3
    # via pip-tools
certifi==2023.11.17
    # via requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via pip-tools
coverage==7.4.1
    # via -r requirements_dev.in
django==4.2.9
    # via
    #   -c requirements.txt
    #   django-stubs
    #   django-stubs-ext
django-stubs[compatible-mypy]==4.2.7
    # via
    #   -r requirements_dev.in
    #   djangorestframework-stubs
django-stubs-ext==4.2.7
    # via django-stubs
djangorestframework-stubs[compatible-mypy]==3.14.5
    # via -r requirements_dev.in
idna==3.6
    # via requests
markdown==3.5.2
    # via -r requirements_dev.in
mypy==1.7.1
    # via
    #   django-stubs
    #   djangorestframework-stubs
mypy-extensions==1.0.0
    # via mypy
packaging==23.2
    # via build
pip-tools==7.3.0
    # via -r requirements_dev.in
pyproject-hooks==1.0.0
    # via build
requests==2.31.0
    # via djangorestframework-stubs
sqlparse==0.4.4
    # via
    #   -c requirements.txt
    #   django
types-pytz==2023.4.0.20240130
    # via django-stubs
types-pyyaml==6.0.12.12
    # via
    #   django-stubs
    #   djangorestframework-stubs
types-requests==2.31.0.20240125
    # via djangorestframework-stubs
typing-extensions==4.9.0
    # via
    #   django-stubs
    #   django-stubs-ext
    #   djangorestframework-stubs
    #   mypy
urllib3==2.1.0
    # via
    #   requests
    #   types-requests
wheel==0.42.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
